Agenda Repository & Agenda task

+++ src/main/java/edu/icet/repository/AgendaRepository.java
@@ -3,6 +3,6 @@ package edu.icet.repository;
 import edu.icet.entity.AgendaEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AgendaRepository extends JpaRepository<AgendaEntity,Long> {
+public interface AgendaRepository extends JpaRepository<AgendaEntity,Integer> {
 
 }
